+++ mobilitydb/src/h3/th3index_latlng.c
@@ -58,7 +58,7 @@ Datum
 Tgeompoint_to_th3index(PG_FUNCTION_ARGS)
 {
   Temporal *temp = PG_GETARG_TEMPORAL_P(0);
-  int32 resolution = PG_GETARG_INT32(1);
+  uint32_t resolution = (uint32_t) PG_GETARG_INT32(1);
   Temporal *result = tgeompoint_to_th3index(temp, resolution);
   PG_FREE_IF_COPY(temp, 0);
   PG_RETURN_TEMPORAL_P(result);
@@ -76,7 +76,7 @@ Datum
 Tgeogpoint_to_th3index(PG_FUNCTION_ARGS)
 {
   Temporal *temp = PG_GETARG_TEMPORAL_P(0);
-  int32 resolution = PG_GETARG_INT32(1);
+  uint32_t resolution = (uint32_t) PG_GETARG_INT32(1);
   Temporal *result = tgeogpoint_to_th3index(temp, resolution);
   PG_FREE_IF_COPY(temp, 0);
   PG_RETURN_TEMPORAL_P(result);
